Replaced some longs with int32s.
git-svn-id: svn://svn.code.sf.net/p/opensludge/code@257 27a91dd4-ec73-4ff0-838b-64dfa05f697e

+++ source/Compiler/moreio.cpp
@@ -1,5 +1,6 @@
 //#include <windows.h>
 #include <stdio.h>
+#include <stdint.h>
 #include <string.h>
 //#include <io.h>
 
@@ -27,13 +28,13 @@ void put2bytes (unsigned int numtoput, FILE * fp) {
 	fputc ((unsigned char) (numtoput % 256), fp);
 }
 
-void put4bytes (long int i, FILE * fp) {
-	fwrite (&i, sizeof (long int), 1, fp);
+void put4bytes (int32_t i, FILE * fp) {
+	fwrite (&i, sizeof (int32_t), 1, fp);
 }
 
-long get4bytes (FILE * fp) {
-	long f;
-	fread (& f, sizeof (long), 1, fp);
+int32_t get4bytes (FILE * fp) {
+	int32_t f;
+	fread (& f, sizeof (int32_t), 1, fp);
 	return f;
 }
