+++ source/ProjectManager/moreio.h
@@ -2,13 +2,14 @@
 extern "C" {
 #endif
 #include <stdio.h>
+#include <stdint.h>
 
 // Input
 
 short getSigned (FILE * fp);
 unsigned int get2bytes (FILE * fp);
 int get2bytesReverse (FILE * fp);
-long get4bytes (FILE * fp);
+int32_t get4bytes (FILE * fp);
 char * readString (FILE * fp);
 char * readText (FILE * fp);
 char * grabWholeFile (char * theName);
@@ -19,7 +20,7 @@ char * grabWholeFile (char * theName);
 void putSigned (short, FILE * fp);
 void put2bytes (unsigned int numtoput, FILE * fp);
 void put2bytesR (int numtoput, FILE * fp);
-void put4bytes (long int i, FILE * fp);
+void put4bytes (int32_t int i, FILE * fp);
 void writeString (const char * txt, FILE * fp);
 void putFloat (float f, FILE * fp);
